refactor(clothes): move ClothesService to services directory and update imports
